Support_Chatbot/manageGPTRequest/requirements.txt
Flask==3.0.2
openai==1.13.3
Werkzeug==3.0.1
python-dotenv==1.0.1
azure-core==1.30.1
azure-ai-ml==1.13.0
azure-search-documents==11.6.0b2
langchain-community==0.0.28
openai==1.13.3
Flask-Cors==4.0.0
langchain==0.1.12